// src/api-client/models/api-response-page-response-branch-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PageResponseBranchResponse } from './page-response-branch-response';

/**
 * 
 * @export
 * @interface ApiResponsePageResponseBranchResponse
 */
export interface ApiResponsePageResponseBranchResponse {
    /**
     * 
     * @type {number}
     * @memberof ApiResponsePageResponseBranchResponse
     */
    'code'?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiResponsePageResponseBranchResponse
     */
    'message'?: string;
    /**
     * 
     * @type {PageResponseBranchResponse}
     * @memberof ApiResponsePageResponseBranchResponse
     */
    'result'?: PageResponseBranchResponse;
}